requirements.txt
python-telegram-bot[ext]>=20.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.27
pytz>=2023.3
pytest>=7.0
pytest-asyncio>=0.21
